Removed dos2unix
dos2unix was running every time to convert the wg0.conf from CRLF (windows) to LF (*nix). VPN providers tend to handle that for you when you download the file in the first place, so that's not necessary. Also, it seemed to be resetting the permissions on the file, generating a message about the file being world accessible.