Update start.sh
Still finding problems from #28.

The eval calls are really causing havoc with curl command parsing. What I'm seeing is that even when getting a valid response from the server while using the setPreferences API, it's not actually updating the port.

So, I took the eval calls out and reformatted the curl calls to use --header where appropriate for the CloudFlare header data.

@tahmidul612, if you get some time, test this one out and see if it works for your use case without the eval calls. If not, we'll need to figure something else out.